cascadya-edge-os-images/ansible/roles/edge-agent/files/src/requirements.txt
annotated-types==0.7.0
paho-mqtt==2.1.0
pydantic==2.12.5
pydantic_core==2.41.5
pymodbus==3.11.4
tenacity==9.1.2
typing-inspection==0.4.2
typing_extensions==4.15.0
